Move View All in Asset Catalog
